# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comments API",
        "description": "API для управления комментариями к новостям и видео постам",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/video-posts": {
            "get": {
                "tags": [
                    "Video Posts"
                ],
                "summary": "Get all video posts",
                "operationId": "48114f1dc6b3d7abd44cd57d44846b00",
                "responses": {
                    "200": {
                        "description": "List of video posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VideoPost"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Video Posts"
                ],
                "summary": "Create video post",
                "operationId": "139fb359061da87c315494d1d97d9f23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Video Post Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Video Post Description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video post created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VideoPost"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/video-posts/{id}": {
            "get": {
                "tags": [
                    "Video Posts"
                ],
                "summary": "Get video post by ID with comments",
                "operationId": "79f98c6e8b6e8f86376a0c3dc6051487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Cursor for pagination",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video post details with comments"
                    },
                    "404": {
                        "description": "Video post not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Video Posts"
                ],
                "summary": "Update video post",
                "operationId": "ae715f8e0f4799227a56d26f09382a70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video post updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VideoPost"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video post not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Video Posts"
                ],
                "summary": "Delete video post",
                "operationId": "4ce420c1fe0d8b4f25bf41cb02adf5cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video post deleted"
                    },
                    "404": {
                        "description": "Video post not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "VideoPost": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Video title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Video description"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}